requirements.txt
rospy
numpy
argparse
pyyaml